fix: Resolve remaining analyzer errors (type conversions and icon)
- spending_evaluation_service.dart: Convert BudgetCategory to String (category.name) and use displayName in recommendations
- dashboard_provider.dart: Add .toInt() conversions for goal amount calculations
- category_breakdown_card.dart: Add .toDouble() for percentage calculation
- goals_progress_card.dart: Add type cast to double for progress calculation
- quick_actions_card.dart: Replace undefined Icons.target with Icons.track_changes

These changes resolve all 5 critical analyzer errors while maintaining functionality.

+++ lib/features/dashboard/presentation/widgets/goals_progress_card.dart
@@ -59,7 +59,7 @@ class GoalsProgressCard extends StatelessWidget {
 
   Widget _buildGoalTile(BuildContext context, FinancialGoal goal) {
     final theme = Theme.of(context);
-    final progressPercent = (goal.currentAmount / goal.targetAmount).clamp(0, 1);
+    final progressPercent = ((goal.currentAmount / goal.targetAmount).clamp(0, 1) as double);
     final isCompleted = progressPercent >= 1.0;
 
     return Column(
